+++ config/webpack.config.babel.js
@@ -198,6 +198,8 @@ export default {
   output: {
     filename: production ? '[name].[chunkhash].js' : '[name].js',
     chunkFilename: production ? '[name].[chunkhash].js' : '[name].js',
+    hashDigest: 'hex',
+    hashDigestLength: 20,
     path: path.resolve(context, PUBLIC_DIR),
     crossOriginLoading: 'anonymous',
     publicPath: ASSETS_ROOT,
@@ -230,7 +232,7 @@ export default {
         test: /\.(woff|woff2|ttf|eot|jpg|jpeg|png|svg)$/i,
         loader: 'file-loader',
         options: {
-          name: '[name].[hash].[ext]',
+          name: '[name].[hash:hex:20].[ext]',
         },
       },
       styleConfig,