+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For details about compatibility between different releases, see the **Commitment
 - Support for resetting end device session context and MAC state in the Console.
 - The Content-Security-Policy header (that was previously behind the `webui.csp` feature flag) is now enabled by default.
 - Default `Cache-Control: no-store` headers.
+- `Cache-Control: public, max-age=604800, immutable` headers for hashed static files.
 
 ### Changed